process_kill

+++ Deploy/TestSetup.ps1
@@ -35,10 +35,14 @@ $JumpCloudConnectKey = $TestOrgConnectKey
 
 #Prechecks
 #check if installer is stuck running and kill
-$process = get-process JumpCloudInstaller
-$process2 = get-process JumpCloudInstaller.tmp
-$process.kill()
-$process2.kill()
+$process = get-process JumpCloudInstaller >$null 2>&1
+$process2 = get-process JumpCloudInstaller.tmp >$null 2>&1
+if (![System.String]::IsNullOrEmpty($process)){
+    $process.kill()
+}
+if (![System.String]::IsNullOrEmpty($process2)){
+    $process2.kill()
+}
 #Clear Temp\JCADMU folder
 if ((Test-Path 'C:\Windows\Temp\JCADMU') -eq $true){
     remove-item -Path 'C:\windows\Temp\JCADMU' -Force -Recurse
